+++ miracle/data/tasks.py
@@ -1,8 +1,4 @@
 from miracle.config import TESTING
-from miracle.data.delete import (
-    DeleteUrls,
-    DeleteUser,
-)
 from miracle.data.upload import Upload
 from miracle.worker.app import celery_app
 from miracle.worker.task import BaseTask
@@ -17,16 +13,6 @@ if TESTING:
     @celery_app.task(base=BaseTask, bind=True, queue='celery_default')
     def error(self, value):
         raise ValueError(value)
-
-
-@celery_app.task(base=BaseTask, bind=True, queue='celery_delete')
-def delete_urls(self, url_ids):
-    return DeleteUrls(self)(url_ids)
-
-
-@celery_app.task(base=BaseTask, bind=True, queue='celery_delete')
-def delete(self, user):
-    return DeleteUser(self)(user, delete_urls)
 
 
 @celery_app.task(base=BaseTask, bind=True, queue='celery_upload')