+++ miracle/web/views.py
@@ -72,9 +72,7 @@ User-agent: *
 Disallow: /__heartbeat__
 Disallow: /__lbheartbeat__
 Disallow: /__version__
-Disallow: /v1/delete
 Disallow: /v1/jwk
-Disallow: /v1/stats
 Disallow: /v1/upload
 '''