+++ miracle/worker/app.py
@@ -20,7 +20,6 @@ from miracle.log import (
 
 TASK_QUEUES = (
     Queue('celery_default', routing_key='celery_default'),
-    Queue('celery_delete', routing_key='celery_delete'),
     Queue('celery_upload', routing_key='celery_upload'),
 )